fix data segment

+++ data_segment.c
@@ -1,19 +1,19 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "logger.h"
-int* dataSegment = NULL;
+unsigned short* dataSegment = NULL;
 int dataSegmentSize = 0;
 int DC = 0;
 
 
-int* get_data_segment()
+unsigned short* get_data_segment()
 {
     return dataSegment;
 }
 
 void init_data_segment(int size)
 {
-    dataSegment = (int*)malloc(sizeof(int) * size);
+    dataSegment = (unsigned short*)malloc(sizeof(unsigned short) * size);
 }
 
 int get_DC()
@@ -26,14 +26,14 @@ int get_data_segment_size()
     return dataSegmentSize;
 }
 
-void dataSegment_add_data(int value)
+void dataSegment_add_data(unsigned short value)
 {
     logger(DEBUG,"dataSegment:: Adding data: %d\n",value);
     if (dataSegment == NULL)
     {
         init_data_segment(1);
     } else {
-        dataSegment = (int*)realloc(dataSegment, sizeof(int) * (DC+1));
+        dataSegment = (unsigned short*)realloc(dataSegment, sizeof(unsigned short) * (DC+1));
     }
     dataSegment[DC] = value;
     DC++;
