+++ data_segment.h
@@ -2,4 +2,4 @@ void dataSegment_add_data(int value);
 int* init_data_segment();
 int get_DC();
 int get_data_segment_size();
-int* get_data_segment();
+unsigned short* get_data_segment();
